requirements.txt
mysql-connector-python==8.0.31
numpy==1.22.4
imutils==0.5.4
dlib==19.24.0
opencv-python==4.6.0.66
matplotlib==3.5.2
requests==2.28.1
Pillow==9.1.1

